refactor(rh): collapse duplicated empleados pages into EmpleadosModule (Sprint 6)
Audit item A3 — three near-identical pages (rdb/dilesa/global) at ~1,284 LOC
collapsed into a single parametrized `components/rh/empleados-module.tsx` and
three thin wrappers totaling 56 LOC. Behavior is preserved exactly, with
per-empresa variations exposed via props (variant columns, filter chip, Sheet
vs Dialog create UI, scope mode).
